requirements.txt
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.19.1
openpyxl==3.0.4
pandas==1.1.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
selenium==3.141.0
six==1.15.0
urllib3==1.25.10
xlrd==1.2.0
